refactor: create notification dropdown component
